feat: externalize pp-display library
Both the search view and the goal view make use of a pp-display library.
Currently only the goal view has an up to date one. The goal is to
make it into a separate and re-usable library that we will then import
in both webviews.

closes issue #838
