+++ client/package.json
@@ -740,8 +740,8 @@
     ]
   },
   "scripts": {
-    "install:all": "yarn install && cd goal-view-ui && yarn install && cd ../search-ui && yarn install",
-    "build:all": "cd goal-view-ui && yarn run build && cd ../search-ui && yarn run build && cd .. && yarn compile",
+    "install:all": "yarn install && cd pp-display && yarn install && cd ../goal-view-ui && yarn install && cd ../search-ui && yarn install",
+    "build:all": "cd pp-display && yarn run build && cd ../goal-view-ui && yarn run build && cd ../search-ui && yarn run build && cd .. && yarn compile",
     "start:goal-view-ui": "cd goal-view-ui && yarn run start",
     "build:goal-view-ui": "cd goal-view-ui && yarn run build",
     "build:dev:goal-view-ui": "cd goal-view-ui && yarn run build:dev",
