+++ client/pp-display/.gitignore
@@ -1,0 +1,7 @@
+node_modules
+.DS_Store
+dist
+dist-ssr
+build
+build-ssr
+*.local
