+++ client/pp-display/package.json
@@ -1,6 +1,13 @@
 {
-  "name": "goal-view-ui",
+  "name": "pp-display",
   "version": "0.0.1",
+  "type": "module",
+  "files": [
+    "dist"
+  ],
+  "main": "./dist/pp-display.es.js",
+  "types": "./dist/main.d.ts",
+  "description": "A react library to format and display Coq Pp types.",
   "private": true,
   "scripts": {
     "start": "vite",
@@ -8,12 +15,15 @@
     "build:dev": "tsc && vite build -m development",
     "preview": "vite preview"
   },
+  "author": "rtetley",
+  "license": "MIT",
   "dependencies": {
+    "@react-hook/resize-observer": "^1.2.6",
     "@vscode/codicons": "^0.0.32",
     "@vscode/webview-ui-toolkit": "^1.2.1",
-    "pp-display": "link:../pp-display",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-hook": "^0.0.1",
     "react-icons": "^4.7.1",
     "uuidv4": "^6.2.13"
   },
@@ -23,6 +33,9 @@
     "@types/vscode-webview": "^1.57.0",
     "@vitejs/plugin-react": "^1.0.7",
     "typescript": "^4.4.4",
-    "vite": "^2.9.18"
+    "vite": "^2.9.18",
+    "vite-plugin-dts": "^4.0.1",
+    "vite-plugin-externalize-deps": "^0.8.0",
+    "vite-plugin-lib-inject-css": "^2.1.1"
   }
 }
