+++ client/pp-display/src/main.tsx
@@ -1,0 +1,4 @@
+import PpDisplay from './pp';
+import {PpString, PpMode, PpTag} from './types';
+export {PpDisplay};
+export type {PpString, PpMode, PpTag};
