+++ client/tsconfig.json
@@ -13,5 +13,5 @@
 		// "noFallthroughCasesInSwitch": true, /* Report errors for fallthrough cases in switch statement. */
 		// "noUnusedParameters": true,  /* Report errors on unused parameters. */
 	},
-    "exclude": ["node_modules", ".vscode-test", "goal-view-ui", "search-ui"]
+    "exclude": ["node_modules", ".vscode-test", "goal-view-ui", "search-ui", "pp-display"]
 }
